kaisensdata2/requirements.txt
djongo
djangorestframework
django
django-cors-headers